pr6 changes round 2
